review: move server tests
